Fix more warnings

+++ src/core/common/message.cpp
@@ -211,7 +211,7 @@ ThreadError Message::MoveOffset(int aDelta)
     assert(GetOffset() + aDelta <= GetLength());
     VerifyOrExit(GetOffset() + aDelta <= GetLength(), error = kThreadError_InvalidArgs);
 
-    mInfo.mOffset += aDelta;
+    mInfo.mOffset += static_cast<int16_t>(aDelta);
     assert(mInfo.mOffset <= GetLength());
 
 exit:
